Exploratory pass at time offsetting with ATL
The idea here is to allow creators to specify the "entrypoint" of their
ATL transforms, such that an animation may appears to be already in
progress when first shown, rather than always at the start of it's run.
The primary use case for this is for repeating sequences that could be
benefit from being offset (statically or dynamically) from one another
(think particle wave) and things like environmental loops (leaves
blowing the wind for example) that can begin to loop unnatural when
always starting the same place.